goexec: cleaning up code; adding documentation.
